Fix crcing of BigInts.
What happened here, is that BigInt dumped all of its data to the end of
the buffer that we provided it with, then we hashed the first part of
it. So, all integers were in fact zero as far as it was concerned.

Fix by actually identifying which bytes we need to hash.

I swear I ran into this in the past and fixed it...