Handle unions in structs accessed with dyn offs.
Previously we just treated them like scalars, which exploded, and was
bad. Instead, actually produce an access to a union field that
represents what we're doing.

This is frequently going to be wrong, because the union situation still
isn't all tied down. However it's going to be right sometimes, as
opposed to right never, so ship it.