Fix k-induction's symbol resolution.
Whenever we have a failed symbol (or any new symbol is added), it ends
up being added to not-the-current-context, because k-induction has many
of them. Boom.

This is highly obnoxious and bad, however given that migrating
expressions occurs across hundreds of call sites and the entire code
base, and this is the simplest fix.