Work around a TACAS14 problem.
Specifically, in the seq-mthreaded dir, we get some kind of implication
logic that has an int instead of a bool type. And that's because the
command line has "-D_Bool=int" in it, like the rest of the competition
benchmarks.

Work around this with this current patch, inserting bool<=>int casts
where it crashes.