Fix a problem with ptr_obj being applied to ints.
Specifically, 01_cbmc_Malloc25 doesn't include stdlib.h, so the ANSI-C
frontend correctly assumes that malloc returns an int. And then in the
body of symex_malloc, we apply pointer_object to the return variable,
which is an integer. Which explodes when we hit the SMT layer.

Fix this by using the un-typecasted address of the allocated object,
which is guarenteed to have pointer type.