Revert "Also rename array size expressions."
This reverts commit 8d0ec5b048d7f3b10ce04662960cf46bd3b647d4.

I'm encountering a serious problem in the 02_eureka02 test: an array
which has ever-so-slightly nondet type. It's size is a symbol, so it
counts as a C99 variable sized array (even though that variable is
constant through the program).

The problem comes when, at some point, the array is accessed with its
size renamed to its constant value (8). In that case, a bounded array
ast is created, which is correct. And then later, an un-renamed array
symbol is used, considers it an unbounded array, and then bursts into
flames.

Fix this by just not renaming array sizes. Life is too short (right now)
to hunt down where this inconsistent renaming is coming from.