Enable bool-in-arrays fixup.
Because we're only using smt_sort to store data about our sorts,
unfortunately we miss cases where the array range is a bool. No bv size
is encoded in smt_sort (because it's a bool, not a bv), so crazy things
occur.

Get around this by enabling the option for all bool arrays to become bit
arrays.