Re-enable Z3 tuples and overflow foo.
This was disabled in the past as I wanted to test that our tuple-beater
and overflow arith was actually working, and Z3 was the only solver
implemented at the time.

Now we have a bunch more solvers, so Z3 can go back to using it's
natives types, which are probably way better.