Take special care when creating array types.
What's occurred here is that there are two different ways of creating an
array sort: the smt codepath, and the as-part-of-a-struct code path. The
difference between the two is that in the former we reduce the domain
sort width down to the smallest size that can accomodate the array size.
While in the latter we don't.

This then leads to type mismatches when we a) try to use the field and
b) try to read the model out of Z3; See 33_s3_srvr_blast_04 before this
commit.