Disable a check on bitvector widths.
Massively irritatingly, C allows for zero length bitfields (for no
reason), and they crop up in the linux based tacas benchmarks. Never
mind the fact that they end up creating structs with no members, which
is undefined behaviour.

Anyway, this getting in the way too much, so remove it for the moment.
No workaround is obvious.