Allow for internal state dumping from dereferences
This new mode tells build_reference_to to collect all the relevant data
about the reference it was about to build, put it in a list, but not
actually build the reference. Then at the end of the call to dereference
we hand all this data back to the caller, via the dereference callback.

The idea here is that things like realloc, and possible some of the
memset/memcpy like functions, can pull the value set out of the
dereference state, with all the additional data that they need to do
whatever it is that they're doing.