Hook the realloc function call.
The CBMC way of doing this would be to create a sideeffect irep that had
realloc baked into this in some way. On the whole, it would be better if
realloc were hooked out during the goto code conversion rather than in
the preprocessor, but I don't think there's a huge amount lost there for
the moment.