Take a shot at reallocating arrays.
This just works by performing a symex malloc to create a new thing of
its type (with a bogus size for the moment), then assign the old data to
the newly allocated hunk.

This invovles changing symex_malloc to return the thing that it's
allocated, rather than just making the assignment at the SSA/valueset
layer.

This still doesn't free the source hunk.