If realloc derefs to nil, leave return val free
That is, if it doesn't point at anything valid, there will be zero
things in the results list. Which we can't build a result from.

However the symex_free code should have baked in some assertions that
the operand isn't invalid, null, and so forth, so we should be able to
happily leave the return value of realloc free and undefined, and those
assertions will the fire when the formula is checked.