Attempt to make vector realloc the vector data.
This means we have to in initially allocate it in the constructors; and
on calls to reserve and so forth, reallocate it.

A problem is going to be that various parts of the vector code put items
into the array before then munging the size, which is the wrong order,
and will trigger various bounds checking bugs.