Group pointer reallocs by type.
That is, if we're reallocating a pointer that has two possible values,
both of which have the same type (which should be a common event...),
then only perform one new allocation and produce an ite chain just for
new hunks value, not an ite chain of all-new hunks.

This means that we don't exponentially increase the number of dynamic
objects that have been allocated whenever we realloc.