Eliminate ieee floating point support.
All this does is serve to confuse: we don't support it in any way right
now, and given that _no-one_ _cares_ about it, we won't for the
forseeable future. This commit can be reverted if we ever change our
minds, but hopefully teh code base will have changed significantly in
the meantime (and more freely without this cruft).