Remove templating of goto_programt.
This was a massive pain in the neck: every time you did something wrong
with some type derived from the goto program type, the magic gcc
diagnostic pixies would through huge reams of template status messages
at you til you drowned.

The only actual achievement of this template was to make the 'guard' and
'code' types in instructiont templated. Perhaps at some point this was
intended to serve as a container for different kinds of instruction and
so forth, but AFAIK it's never been used for anything else, and it can
quite happily be discarded right now. The result will be better
compilation time, fewer templates, and better messages.