Revert current realloc implementation.
Performance is really suffering here. Looking at some of the traces from
the esbmc-vector tests, the issue is that when a vector gets inserted
into, the reallocation is conditional. That is, if the capacity is
insuffient, then a realloc occurs, but not if there's enough space. That
can't be constant propagated / simplified away, so both happen, and we
thus end up with a new vector array every time we modify it.

That then results in a large switch/case expression every time we try to
access the vector, which presumably is what's killing performance. It
could also be that the sheer number of addresses taken is exploding the
number of address-space constraints, but as nothing casts to an int, I
don't think it can truly be contributing a lot.