Work around some invalid formulas from derefing.
What happens here: in tests souch as esbmc-cpp/ch21_8, we have an
integer access into a multimap strucutre, at a nondeterminstic offset,
so we try to represent all possible accesses. Which is actually pretty
simple because there are only three fields.

However, in that test the _value array contains a bunch of double's,
which aren't compatible with the integer type we're trying to get out of
them. In this case we were generating a byte extract: however that only
ever results in a byte. When Z3 encounters this, it croaks.

Fix this by returning a free value instead. In the case of ch21_8, the
access to _value as an integer is never actually used because the
pointer never points that far out (which is good). And even if it was,
the bit representation of doubles is machine specific, so returning an
arbitary value is fine IMO.

In other cases, this might break /some/ things, but they're _highly_
likely to be invalid accesses. The exception is bytes: you can always
byte extract a value out there.