Don't assume 8 bit operands to concat exprs.
This was fine in the past, where we were always writing values into 8
bit arrays. However what we have now is (in esbmc-cpp/ch21_9) the
(nondeterministic) assignment of a 64 bit fixedbv into a 32 bit array.
(The guard is never true, but it's a possibility).

Handle this by having the concat code being agnostic about the byte
sizes of what it's working with. Note that it's mostly assuming little
endian semantics.