Eliminate comma operator that made its way to SMT
There are two bugs here: first is that the author of operator<< wasn't
aware that the first part of his test (i < NUM_SIZE) would be discarded
due to the way the comma operator works. Secondly, that the operator
wasn't eliminated as a side-effect: there's a sequence point in the
middle of it.