Attempt to avoid allocating negatively sized str.
The problem here is that if we allocate say, -4, then the whole of the
address space gets used up and the solver can't allocate where things
are. And this is bad.

This is a hard problem to solve given that the formula goes unsat the
moment that this allocation /might/ occur. For the moment, this patch
fixed a few tests (such as string-error-construct-2).

I'm not certain how this was working on the cpp branch, may have been to
do with slicing :/