Fix migration namespace for parallel k-induction.
Due to the irep2 guff requiring us to be able to move back and forth
between old-irep and new-irep, the program namespace is required to look
up symbols, and work out what level of renaming has been applied to a
symbol. This needs to be manually munged for parallel k-induction.