Massage the k-induction IPC method.
There are a few issues here:
  * Having the 'read' end of the pipe non-blocking means the main esbmc
    instance busy waits
  * Reading into a huge array of results needlessly complicates things
    when one item can be read at a time. This also means we can never
    detect short reads
  * Iterating over the huge array of results will read stale results

To solve this, don't put the pipe into non-blocking mode, and only read
one result from it at a time. This is a) safe against short reads, and
b) much more efficient.

Using only one result at a time requires that the result interpretation
code doesn't iterate over an array of results; this looks messy in the
patch, but is fairly straightforward.