Store concrete template types into template scope.
Right now, when one wants to look up what type a template parameter has
in a template, digging around in the 'template_map' has to occur, to
find the mapping from the parameter name to the concrete type.
Unfortunately, that template map is unavailable outside of the templates
instantiation. This means that the instantiation of nested templates
breaks, because it can't work out what the outer templates types were
when /it/ is instantiated.

To fix this, instead try installing the parameter names and their
concrete types into the scope of the instantiated template. This is safe
because it's only applying to the instantiated scope, not the general
templates scope. It also exists for the whole lifetime of the C++
conversion/typechecking, meaning that we can find this information at a
later date, not just when instantiating a template.

This may make the template map redundant. We shall see.