Resolve instantiated type args from several srcs
This way, we can resolve types that were inserted into the template_map
as part of guessing what to resolve a template to; as well as from the
template scope if we're accessing the type /after/ instantiation has
occurred. i.e. for nested templates.

A small program with a templated constructor now successfully parses.