Make new template arg resolver cope with non-types
That is; if we have a template argument that's a constant value (i.e. an
integer, a boolean, or the contents of gcc-template-tests/arg6), then we
fail to correctly bake that value into the symbol table.

So, test to see whether we're dealing with a type or a value when
putting template arguments into the symbol table, and act appropriately.