Implement SFINAE.
"Substitution failure is not an error" is a behaviour where, if a
template is considered as a candidate in a overloaded function
resolution, but not picked, then errors during instantiation shouldn't
be reported as an error.

Currently whenever we consider a template for a function overload, we
also instantiate it. And that instantiation ends up producing an error
when it's typechecked later. Which is erronous.

To work around this, whenever we guess a function instantiation during
overload resolution, mark it as being a speculatively instantiated
template function. Then, if we pick it, mark it as being in use.
Finally, when we start looking at a function to be typechecked, bail out
if it was a speculative instantiation that wasn't used. This means that
it doesn't ever get typechecked, which meets the requirements of SFINAE
(although we don't actually run into any substitution errors).