Put all types into the 'cpp' namespace.
If you look at the 'esbmc-cpp/bicycle' test right now, it won't compile,
because of a call to gettimeofday. In that location, we take the address
of a `struct timeval`, which ends up being cpp::tag.timeval, and call
gettimeofday, which expects a pointer to struct named c::tag.timeval.
Those two pointers don't compare equal, because one of those names
doesn't exist. Thus the code is spat out as an invalid implicit
conversion.

To fix this, just ditch any types in the c:: namespace, and put
everything into the cpp:: namespace. This is fine: so long as all code
that tries to look up a (fully qualified) type or name it is consistent.