Avoid converting failed template members
The current way that SFINAE works is to instantiate a template, but if
it's unused, then to ignore that function when it gets typechecked.
Unfortunately that still turns up in the class type, which will have
been baked into the code of many other ireps.

This only really manifests itself when the type gets converted to irep2,
when all sorts of weird and wonderful (and massively invalid) type
symbols start turning up. This is exactly the kind of thing that's
supposed to happen (i.e.: irep2 breaks when you hand it broken code),
and what's really wrong is that these bad templates are ending up baked
into class definitions.

To work around this, when converting a component member of a struct
type, look up its name: if the identifier is in fact an unused
instantiation of a template member, then discard it.