Avoid putting template args into a scope twice.
This occurs when we have a declaration of a template, but not a
definition of it. We might end up "instantiating" it without a body
several times, then filling in the gaps when a definition turns up.

In these cases though, don't put the template args into the scope more
than once (because that fails). Instead, detect that we've already
instantiated it (partially or fully), and don't make the function call
to put the args in scope. This hopefully avoids overcomplicating the
args-in-scope logic.