Match method template args against operands.
Previously in this context, the fargs.operands array had three operands
(one of which was the implicit 'this' operand), and the function
template only had two, because it hadn't been instantiated yet. This
then lead to crazy when template argument deduction was applied.

Fix this by detecting that condition, and indexing the operand and
argument arrays to take this into account.