Don't special-case non-member function invocations
That is; when called in the context of a method invocation, it seems we
deleted the object and tried to match normal function templates too.

Which is crazy; don't do this, thus fixing template method invocation.
Template methods aren't class members at this point, so they were
previously going down this path.