Fix method-templates hiding other identifiers.
Take the 'pair' class as an example (in library/utilities) -- it has a
templated constructor, which has the name 'pair'.

When we instantiate that template constructor (within the 'pair'
template itself), we make an inner scope to represent that template. We
also put the 'pair' definition into that scope for reasons that aren't
entirely clear.

The problem is that inside the 'pair' method definition, we try to look
up the 'pair' template again (for pair<U,V>). And what we get back is
the function template, because that's in the closest scope, and is the
only thing there, so hides 'pair' in the higher scope. That then leads
to the function template being instantiated as a type, which explodes.

Fix this by, whenever we generate a method template scope, recording
that it uses the containing classes scope, using the same infrastructure
that the 'using' keyword does. This means that any lookup for 'pair' in
the template will also look in the class scope, although it won't go any
higher than that.