Disambiguate copy constructor ambiguity.
Classes always have a default copy constructor. However, templated
constructors can be instantiated that overload it exactly, and thus
there's an ambiguity of whether to call the default copy cons or a
template that might match.

To get around this, the C++ spec says that no template is ever
instantiated when calling the copy constructor. Thus, if after all
disambiguation the best candidates are
  a) the default copy constructor, and
  b) some template constructors
Then erase the template constructors. This implements precisely what the
spec wants, as far as I can tell. If another constructor is declared
with default arguments to the extent that it matches the copy
constructor too, then the program is ill formed, and reported as an
ambiguity.

This means adding yet another irep flag, hanging off the return type of
the default copy constructor. This is fairly crap, but in the
circumstances there's nothing else to be done.