Pass 'this' constructor arg for cons. overloading
Right now constructor overloading doesn't entirely work in the presence
of templates. Specifically, picking the right method requires knowing
what the base object is, otherwise a 'foobar<T>' constructor might be
called for a 'foobar<X>' type.

Try to work around this by decorating the function call typechecker with
the 'this' object. I don't think it can be passed down directly as an
argument, because so much of the existing typechecking code assumes that
that doesn't happen. (The argument doesn't become the 'has_object'
object in fargst, and specialised code elsewhere already exists to
detect and add a missing 'this' argument, after name resolution).

The hope here is that by providing the object for name resolution we can
have overloading working correctly; and not need to modify any other
large portions of code.