Fix name resolution of constructors.
Don't pull constructor components out of class definitions -- there may
be templates with the class names in scope that need to be instantiated.
Instead, use the mechanism that C++ already has (name lookup) to
identify the candidate constructors for a class.

(Without this, I don't believe class constructor templates are ever
 instantiated).