Don't disambiguate function templates.
They don't have a 'this' parameter while uninstantiated, and so can't
/actually/ be correctly overloaded til later. So, instantiate all
candidate templates, and let later disambiguation functions resolve
which one to use.

Also special case the creation of constructor expressions -- apparently
the rest of the code base expects these to be 'symbol' expressions
rather than member exprs of the class being constructed.