Resolve a recent ambiguity I introduced.
If we have the following:

template <class X>
class foo
{
    template <class Y>
    foo(const foo &bar, const foo<Y> &baz);
};

Then in the foo constructor, the first foo resolves to the foo class
instantiated with template argument X; and the second foo refers to the
template again and has to be deduced during overloading.

Do this, by a) knowing that this will only ever occur in the context of
such a template, and b) knowing that there will be exactly two symbols,
one a template, the other a class.