Insert method templates into the correct scope.
It needs to be typechecked in its own scope to keep it independant of
other instantiations, but then when it's stored, that has to go into the
class itself.

Alas, this code is ugly, because we can't actually detect from the type
of the method that it's a method template; so we have to look at the
scope context to detect that.