Fix a bunch of overloaded operator return types
It turns out that the postmodification operators are supposed to return
a temporary, not a reference to a temporary. If they return the
reference, we (legitimately) get incorrect behaviour, because the object
has been destroyed. Wheras actually we're supposed to make the call-site
allocate and initialize a temporary for the return.

Also, Huuuurrrrrrr