Detect already-instantiated templates better.
Basically; trying to shoehorn these templates into the name lookup scope
is going to fail, because either we'll end up disambiguating them, or
they'll hide other symbols. So instead, store the list of things that
have been instantiated in a field in the type of the template that's
being instantiated. Then we can just look up the name of the
instantiation signature, and know whether there's already a symbol for
that.

There's some uncertainty about the conditions under which instantiation
continues with incomplete structs / the like, I'll just have to cross
that bridge when I come to it.