Attempt to cope with recursive templates.
That is: templates that refer to themselves with name<args> in their
body. This is sort-of-legitimate I believe; however it leads to
templates attempting to instantiate themselves, which is a poor plan.
Therefore, attempt to work around this by detecting when we're
instantiating a template that we're in the middle of instantiating, and
create a symbolic type that points to where the symbol will eventually
be.

If anything attempts to access itself cyclicly, such as:
  typedef myself<args>::type type;
Then it'll explode with a failure to lookup a symbol name. Too bad, it's
a bad program.

The implementation detail here is that we have to work out what the name
of the struct will be once it's instantiated, and if we detect some
recursion then create a new symbol that points at that. A small amount
of faff also has to occur to make other pieces of code (such as the
member operator) swallow this.

On the whole, relatively ugly, but I imagine this is necessary.