Make double-templ-arg-insertion non-fatal.
So far this has been a brilliant test of whether something is wrong with
the template gunge; however I've now reached a point where it's pretty
much a legitimate problem to occur.

Specifically, if we have an undefined (but declared) template like
char_traits, then it may be referred to (but not used) multiple times
before being defined. In that circumstance, we end up producing an
incomplete struct type and returning that. However, detecting this and
jumping through hoops to work out whether to insert template args, is
actually quite difficult.

To avoid making this code path brain dead, fix this by relaxing the
constraint that we only insert template args into the scope once, and
recover from it happening safely. I imagine this means that any
instances of recursive instantiation will end up printing the 'maximum
recursion depth' error, hopefully.