Handle zero sized C++ classes.
These tend to crop up upon occasion. The particular problem is that,
because they have no storage, they don't actually contain any data, and
so both the model checker and SAT solver will either throw their cookies
at seeing them, or will grow mountains of special-case code.

The cleaner solution (IMO) is to just prevent them from entering the
rest of the model checker, entirely. The only case where we have this
problem (that I'm aware of) at the moment is in assignment of PODs,
which the C++ frontend issues because it can see that no code needs to
be invoked. Detect this condition by looking at the size of the struct
being assigned, and if it's zero, perform no assignment.

It's safe to assume that this is OK, because anything that has no
storage size has no data, and so has no effect on the program.