Don't skip over asserts with false guards.
Sometimes this end up being the assert that's causing the satisfying
assignment. It's unclear why this situation exists (where the guard is
false but the assert causes a problem), however it's a) already an
error in that case, and b) easier to fix if I can _see_ the assertion in
question that's causing the trouble.