Refuse implicit string construction from int.
This constructor isn't in the C++ spec for string, but appears to be
useful in the construction of temporary strings that then get more data
filled in. And that's fine, except that this constructor can implicitly
construct a string from an integer, which leads to huge amounts of pain.

To prevent that, declare it explicit, which should prevent that implicit
conversion from being performed.