Enumerate the set of ostream operator<<'s.
The spec has an explicit list of operators that ostream is supposed to
support, and if we try to template this then we end up suffering
overloading rules and other conversion problems. In particular, ch22_2
ends up deciding to convert an 'int' to an internal class instead of
instantiating the operator<< template.

To solve this, don't template, and instead just enumerate the list of
things we're supposed to support shifting out of an ostream. This fixes
the overload resolution problem there.