Add a class for more detailed conversion data.
That is; there are various points where we might have to disambiguate
because of the really intricate rules in C++ (like 13.3.3.2.4), where
casts to void ptr are worse that other kinds of casts, and so forth. We
can only know enough to make that kind of decision by inhaling all the
relevant data about a conversion. This class is going to store it.