Lawnmower conversion sequence scoring.
A large number of ambiguities in overloading are resolved with some
random bits of data about the conversion that was performed. Does it
cast a pointer to bool, how far does it cast a class type towards the
base type, and so forth.

To successfully disambiguate, we need that information to be around to
make a decision. (In the current case, disambiguating a cast from a
pointer to bool, and pointer to void ptr, which is killing ostream
operator<< right now. The latter is the best, FYI).

To achieve this, create a class that can store data about what a
conversion contains, and pass it down to the conversion code. We can
then annotate it will all kinds of flags, then make decisions on which
conversion is best later.

This patch itself shouldn't change any behaviour, although various
esbmc-cpp tests are broken due to the aforementioned ambiguity.