Reduce unwind bound for a test.
This is currently taking more than an hour to complete. There's no point
testing it in that case because we never get a result. So instead,
reduce its bound to something that we /can/ test. If there are errors
deeper, then it's simply infeasible for us to detect them in the course
of this testing.