Resolve templated functions better.
If there's a templated function that has a better rank than other
functions, then we should pick it. However, the 'template_distance' used
to distinguish between similar templates is added to the rank, and
messes with this.

Thus, we should make this into two separate tests, and only check the
template distance when we're deadling with two templates.