Don't change the sym table while iterating over it
As the iterators to it may be invalidated if the thing is rebalanced due
to an insertion, which can happen when functions are typechecked, and
the original author clearly knew because he put an outer retry loop in
this function.

This bug seems to have always been present, I just only triggered it
recently. Fix it by taking a list of symbols to be typechecked, then
typecheck them, so that we're not using an unstable hash map iterator
during typechecking.