Gracefully fail to match complex template args.
If we have:

  template <class foo>
  void bar(std::list<foo> &baz);

And we pass as a parameter some other type, qux, then various things
break. The problem is that the template argument guessing code tries to
match up std::list<foo> with qux, by finding qux's template arguments
and assigning them to 'foo'.

However, because qux doesn't have any, none are found, and the code
looks through the scopes all the way to the root, and ends up looking up
template arguments in the root scope, which croaks.

Fix this by detecting better when we've failed to look up template
arguments, and exiting gracefully from the template guessing, recording
our failure by leaving the template argument unassigned. Doing so is
necessary for SFINAE to work, etc.