Fix taking addr of struct members.
This code wasn't updated from the previous memory model implementation,
and instead of calculating the offset into a structure, was creating a
series of member/index exprs to pull data out of it. And this was then
discarded by the new dereference code.

Fix this by not wrapping symbols in member/index exprs. Instead, compute
their offset, which can then be used to accurately dereference things in
the future.