Fix calculation of array bounds in dereferencing.
The range guard here is designed to limit the expression we're building
to only apply when the index is in the range of the array. That stops
the solver from picking an array element that's out of bounds. (Bound
assertions are someone elses problem).

However for some reason this was mixing units: the array size (in
elements) was being compared to the actual byte offset into the array /
struct (see context). Which is wrong. Fix by adding a multiply, bringing
us back up to byte size.

Note that this ends up with us encoding the range guard twice, once in
this array code, and again in the struct member accessor. This can
probably be improved, but isn't currently harmful.