Hack the size type of the list class.
Right now, the templated insert method matches the call in list_size
better than the pos-count-value form of insert. This is because the
conversion from int to unsigned int counts as a promotion apparently, so
scores worse than the template, which can just be signed int all the
way.

The solution for now is to make the size type of ESBMC lists a signed
int so that it matches exactly. It's very unclear to me why exactly all
other compilers seem to accept a wide range of ints in this position
without trying to match the template.