Correct calc of nondet pointer offset in derefing
Normally the base object we're dereferencing has some kind of pointer
offset, which we pull out with the pointer_offset2t irep. However, there
can be additional offsets introduced by outer expressions, stored in the
scalar step list. An example of this (where it currently breaks) is
cpp test esbmc-map/map_string_class-3.

There, an access to this->str[1] in StringClass::operator= calculates
the pointer offset of 'this', and discards the extra offset introduced
by ->str[1]. As a result, the wrong data gets stored, and the test
fails.

Fix this by pulling the offset out of the scalar steps and baking it
into the nondet offset expression.