Reject deduced templ args with non-exact-matches.
The spec demands that only a few conversions are permittable between
function template arguments and their deduced types. They would seem to
be the set of exact match conversions. So, test for this, and abort
deduction if we've deduced an argument that doesn't exactly match
another argument.

This addresses the special case where:

  template <class foo>
  bar(foo a, foo b);

  bar(baz, qux);

baz and qux have differing types that need promotion or conversion to
become the same type. This will now be disallowed. All other
circumstances where the types aren't an exact match are handled by the
fact that we only ever store the same type into the template map (and
the only conflict that can cause, is this one).