Don't check array bounds involving dereferences.
Otherwise we end up checking the lexical bound of some kind of array,
even though what's actually being accessed is whatever the dereferenced
pointer points at. This is a common thing in the C language, where you
get:

  struct foo {
    unsigned int bar;
    struct baz qux[1];
  };

Which is then allocated with extra space for extra elements in qux. In
that case, if we dereference a foo pointer accessing a qux element more
than zero, the dereference is actually valid (because enough space was
allocated).

Hence, in this patch, we disable the array bounds check in that
scenario.