Massively simplify assigning to concats.
This occurs when we decompose a dereference into a selection out of a
byte array. When we assign to such an expression, that has to be further
decomposed into a series of assignments into that array.

Previously this code attempted to do the right thing based on the size
of each side, assuming it'd take a particular format. This had made it
fairly brittle, alas.