Fix a massive bool-conversion brain fail in Z3.
This was using the Z3 C++ api's operator!=, which encoded a 'distinct'
ast. However, that was quite happy to declare that two things were
distinct based on the fact that their _sorts_ were different, and
nothing else! So when I was comparing a 32 bit int against an 8 bit
value, it was always distinct. Which is crazy.

Fix this by generating the reference zero from the bitwidth of the
operand it's being compared against.