Fix a brain fail in symex goto.
What we're looking for here, is to store the jumping insn so that we can
work out what loop number we're jumping back over. However, this was
storing a reference, so when cur_state->source was updated, our view of
what was happening changed too. Which was bad.

Fix this by taking a copy.