Declare symex_throw::thrown_obj in namespace.
This was leading to problems with phi nodes not being generated due to
some assignments being silently discarded, that then turned into
exceptions after this was fixed in e9f03db5.

Address this problem by correctly putting the thrown_obj variable in the
namespace.

This then raises the issue of what type it should have -- something
that'll have to be massaged in the future.