Delete convert_array_equality, use ast eq method
This totally breaks the SAT array converter. However the only other
choice was maintaining a mix of expressions and ast pieces throughout
the core of all this SMT conversion, which I'm not really willing to do.
This will have to be fixed at a future date. I don't think there was
anything fundemental that the array flattener was actually doing that
required the whole expression.