Go slightly crazy squashing a 0-sized-bv problem.
This is either a problem with infinite arrays being interpreted to
having a domain width of zero, or with the creation of new z3_smt_sorts
having some kind of overloading fail whereby the width/bool constructor
is picked instead of the width/domainwidth one. Could have been both
too.

Either way, it seems fixed now.