Put array_smt_ast elem init in constructor
Avoiding doing it in numerous random places. This seems to make sense at
the moment: all the fields in the array need to start free, wheras
tuples might be constructed using completely deterministic data.

This raises the question of whether assignments of tuple arrays is going
to be a thing?