Link up null/invalid tuple symbols.
Previously these were joined through the symbol table, in that when you
generated a tuple with name 'NULL', then it's values would already be
bound to the values you expect. However, this broke with the
tuples-in-smt-ast's alteration in the previous commit.

To get around this, intercept the case where a tuple symbol is created,
and if it's one of the special case symbol (0, NULL, INVALID), then
return the ast representing that symbol.

I tried to do this via the expression cache, however those values can
have different types in different contexts, so it's infeasible for them
to be matched in that manner.