Fix type mismatches in address-of expressions.
Right now, when one converts the address of a symbol, it might end up
with a different type due to the vagaries of ESBMC and some pointer
arithmetic. This isn't initially an issue, but it means that the
address-of a value might not be in the smt cache, and conversion is
required again.

This used to work because the address of that value had a name, and was
linked up in the symbol table. However now that tuples are just being
stored statically, if something gets converted twice, then it has two
different values.

Ultimately the problem here is that the base expression of the address,
&whatever, can have different types. To get around this, build a
canonical address-of expression in convert_identifier_pointer, and store
that in the cache instead (and check it whenever an address has it's
address taken). This leans on the fact that the data object never has a
different type, which is much more reliable than the address-of
expression never having a different pointer type.