Only perform smt tuple assign at top level.
That is, it's only _really_ valid to translate an equality to an
assignment when we know for a fact that an assignment is what was
desired. Fix this by making symex_target_equationt let us know that an
assignment is happening through convert_assign, and do all special
casing in that function.

This makes this much safer, we can always use equalities everywhere
else. We can also shoehorn array-assignment into this too.