Explicitly make long int's int64_t's and the like
On 32 bit machines, where the 'long int' flavour of bv_val is disabled,
if a int64_t is passed to bv_val it has to be converted down to some
other kind of integer. And because all of those candidates require a
conversion, it's ambiguous which method to pick.

Fix this by relying on explicit width integer types, that wonderful part
of C99 that microsoft refuses to use :/