Fix error in for loop compilation, add test
What was happening here, is that iteration statements in for loops that
contained both side-effects _and_ other code, had the other code
discarded. This is quite a serious fault :/

It was made possible by the fact that anyone looking at this piece of
code would have assumed that the function calls in there would have been
parsed as a function call side-effect, wheras actually it was coming
through as a 'function_call' statement. This is why having two ways of
representing things is _bad_.